XMB 1.9.11 - Bugs - Security Issue Nos. 434 through 436.
git-svn-id: https://svn.code.sf.net/p/xmb-forum/code/xmb19x/trunk/source@2489 4328ac3d-c443-4ce5-a636-224d0ca14295

+++ cp.php
@@ -420,6 +420,7 @@ if ($action == "settings") {
         <a href="#7"><?php echo $lang['admin_main_settings7']; ?></a><br />
         </span>
         <form method="post" action="cp.php?action=settings">
+        <input type="hidden" name="token" value="<?php echo nonce_create('mainsettings'); ?>" />
         <table cellspacing="0" cellpadding="0" border="0" width="<?php echo $tablewidth?>" align="center">
         <tr>
         <td bgcolor="<?php echo $bordercolor?>">
@@ -624,6 +625,8 @@ if ($action == "settings") {
         </tr>
         <?php
     } else {
+        request_secure('mainsettings', '', X_NONCE_FORM_EXP, FALSE);
+
         $sitenamenew = postedVar('sitenamenew');
         $bbnamenew = postedVar('bbnamenew');
         $siteurlnew = postedVar('siteurlnew');
@@ -858,6 +861,7 @@ if ($action == 'rename') {
     }
 
     if (onSubmit('renamesubmit')) {
+        request_secure('renamemember', '', X_NONCE_AYS_EXP, FALSE);
         $vUserFrom = postedVar('frmUserFrom', '', TRUE, FALSE);
         $vUserTo = postedVar('frmUserTo', '', TRUE, FALSE);
         $adm = new admin();
@@ -868,6 +872,7 @@ if ($action == 'rename') {
         <tr bgcolor="<?php echo $altbg2?>">
         <td>
         <form action="cp.php?action=rename" method="post">
+        <input type="hidden" name="token" value="<?php echo nonce_create('renamemember'); ?>" />
         <table cellspacing="0" cellpadding="0" border="0" width="550" align="center">
         <tr>
         <td bgcolor="<?php echo $bordercolor?>">
@@ -936,6 +941,7 @@ if ($action == 'forum') {
         <tr bgcolor="<?php echo $altbg2?>">
         <td>
         <form method="post" action="cp.php?action=forum">
+        <input type="hidden" name="token" value="<?php echo nonce_create('massedforums'); ?>" />
         <table cellspacing="0" cellpadding="0" border="0" width="90%" align="center">
         <tr>
         <td bgcolor="<?php echo $bordercolor?>">
@@ -1176,10 +1182,12 @@ if ($action == 'forum') {
         </tr>
         <?php
     } else if ($fdetails && noSubmit('forumsubmit')) {
+        $key = template_key('editfid', $fdetails);
         ?>
         <tr bgcolor="<?php echo $altbg2?>">
         <td align="center">
         <form method="post" action="cp.php?action=forum&amp;fdetails=<?php echo $fdetails?>">
+        <input type="hidden" name="token" value="<?php echo nonce_create($key); ?>" />
         <table cellspacing="0" cellpadding="0" border="0" width="100%" align="center">
         <tr>
         <td bgcolor="<?php echo $bordercolor?>">
@@ -1325,6 +1333,7 @@ if ($action == 'forum') {
         </tr>
         <?php
     } else if (onSubmit('forumsubmit') && !$fdetails) {
+        request_secure('massedforums', '', X_NONCE_FORM_EXP, FALSE);
         $queryforum = $db->query("SELECT fid, type, fup FROM ".X_PREFIX."forums WHERE type='forum' OR type='sub'");
         while($forum = $db->fetch_array($queryforum)) {
             $displayorder = formInt('displayorder'.$forum['fid']);
@@ -1424,6 +1433,7 @@ if ($action == 'forum') {
         }
         echo '<tr bgcolor="'.$altbg2.'" class="ctrtablerow"><td>'.$lang['textforumupdate'].'</td></tr>';
     } else {
+        request_secure('editfid', $fdetails, X_NONCE_FORM_EXP, FALSE);
         $namenew = addslashes(htmlspecialchars(postedVar('namenew', 'javascript', FALSE), ENT_COMPAT));  //Forum names are historically double-slashed.  We also have an unusual situation where ENT_COMPAT is the XMB standard.
         $descnew = postedVar('descnew');
         $allowhtmlnew = formYesNo('allowhtmlnew');
@@ -1656,6 +1666,7 @@ if ($action == "members") {
             <tr bgcolor="<?php echo $altbg2?>">
             <td align="center">
             <form method="post" action="cp.php?action=members">
+            <input type="hidden" name="token" value="<?php echo nonce_create('masseditmems'); ?>" />
             <table cellspacing="0" cellpadding="0" border="0" width="91%" align="center">
             <tr>
             <td bgcolor="<?php echo $bordercolor?>">
@@ -1735,7 +1746,7 @@ if ($action == "members") {
                 <?php if (X_SADMIN) { ?>
                 <br /><a href="editprofile.php?user=<?php echo recodeOut($member['username']); ?>"><strong><?php echo $lang['admin_edituseraccount']; ?></strong></a>
                 <?php } ?>
-                <br /><a href="javascript:confirmAction('<?php echo addslashes($lang['confirmDeletePosts']);?>', 'cp.php?action=deleteposts&amp;member=<?php echo recodeJavaOut($member['username'])?>', false);"><strong><?php echo $lang['cp_deleteposts']?></strong></a><?php echo $pending ?>
+                <br /><a href="cp.php?action=deleteposts&amp;member=<?php echo recodeOut($member['username'])?>"><strong><?php echo $lang['cp_deleteposts']?></strong></a><?php echo $pending ?>
                 </td>
                 <td><input type="text" size="12" name="pw<?php echo $member['uid']?>"></td>
                 <td><input type="text" size="3" name="postnum<?php echo $member['uid']?>" value="<?php echo $member['postnum']?>"></td>
@@ -1777,6 +1788,7 @@ if ($action == "members") {
             <?php
         }
     } else if (onSubmit('membersubmit')) {
+        request_secure('masseditmems', '', X_NONCE_FORM_EXP, FALSE);
         $query = $db->query("SELECT uid, username, password, status FROM ".X_PREFIX."members $where");
 
         // Guarantee this request will not remove all Super Administrators.
@@ -1975,35 +1987,49 @@ if ($action == "ipban") {
 }
 
 if ($action == "deleteposts") {
-    require('include/attach-admin.inc.php');
     $member = postedVar('member', '', TRUE, TRUE, FALSE, 'g');
-    $countquery = $db->query("SELECT tid, COUNT(*) AS postcount FROM ".X_PREFIX."posts WHERE author='$member' GROUP BY tid");
-    deleteAttachmentsByUser($member);
-    $db->query("DELETE FROM ".X_PREFIX."posts WHERE author='$member'");
-    $db->query("UPDATE ".X_PREFIX."members SET postnum = 0 WHERE username='$member'");
-    while($threads = $db->fetch_array($countquery)) {
-        $db->query("UPDATE ".X_PREFIX."threads SET replies=replies-{$threads['postcount']} WHERE tid='{$threads['tid']}'");
-    }
+    if (noSubmit('yessubmit')) {
+        $key = template_key('delps', $member);
+        ?>
+        <tr bgcolor="<?php echo $altbg2; ?>" class="ctrtablerow"><td><?php echo $lang['confirmDeletePosts']; ?><br />
+        <form action="cp.php?action=deleteposts&amp;member=<?php echo recodeOut($member); ?>" method="post">
+          <input type="hidden" name="token" value="<?php echo nonce_create($key); ?>" />
+          <input type="submit" name="yessubmit" value="<?php echo $lang['textyes']; ?>" /> -
+          <input type="submit" name="yessubmit" value="<?php echo $lang['textno']; ?>" />
+        </form></td></tr>
+        <?php
+    } elseif ($lang['textyes'] == $yessubmit) {
+        request_secure('delps', $member, X_NONCE_AYS_EXP, FALSE);
+        require('include/attach-admin.inc.php');
+        $countquery = $db->query("SELECT tid, COUNT(*) AS postcount FROM ".X_PREFIX."posts WHERE author='$member' GROUP BY tid");
+        deleteAttachmentsByUser($member);
+        $db->query("DELETE FROM ".X_PREFIX."posts WHERE author='$member'");
+        $db->query("UPDATE ".X_PREFIX."members SET postnum = 0 WHERE username='$member'");
+        while($threads = $db->fetch_array($countquery)) {
+            $db->query("UPDATE ".X_PREFIX."threads SET replies=replies-{$threads['postcount']} WHERE tid='{$threads['tid']}'");
+        }
 
-    // Delete Empty Threads
-    // This will also delete thread redirectors where the redirect's author is $member
-    $tids = array();
-    $movedids = array();
-    $countquery = $db->query("SELECT t.tid FROM ".X_PREFIX."threads AS t LEFT JOIN ".X_PREFIX."posts AS p USING (tid) WHERE t.author='$member' GROUP BY t.tid HAVING COUNT(p.pid) = 0");
-    while($threads = $db->fetch_array($countquery)) {
-        $tids[] = $threads['tid'];
-        $movedids[] = 'moved|'.$threads['tid'];
-    }
-    if (count($tids) > 0) {
-        $tids = implode(', ', $tids);
-        $movedids = implode("', '", $movedids);
-        $db->query("DELETE FROM ".X_PREFIX."threads WHERE tid IN ($tids) OR closed IN ('$movedids')");
-        $db->query("DELETE FROM ".X_PREFIX."favorites WHERE tid IN ($tids)");
-        $db->query("DELETE FROM d, r, v "
-                 . "USING ".X_PREFIX."vote_desc AS d "
-                 . "LEFT JOIN ".X_PREFIX."vote_results AS r ON r.vote_id = d.vote_id "
-                 . "LEFT JOIN ".X_PREFIX."vote_voters AS v  ON v.vote_id = d.vote_id "
-                 . "WHERE d.topic_id IN ($tids)");
+        // Delete Empty Threads
+        // This will also delete thread redirectors where the redirect's author is $member
+        $tids = array();
+        $movedids = array();
+        $countquery = $db->query("SELECT t.tid FROM ".X_PREFIX."threads AS t LEFT JOIN ".X_PREFIX."posts AS p USING (tid) WHERE t.author='$member' GROUP BY t.tid HAVING COUNT(p.pid) = 0");
+        while($threads = $db->fetch_array($countquery)) {
+            $tids[] = $threads['tid'];
+            $movedids[] = 'moved|'.$threads['tid'];
+        }
+        if (count($tids) > 0) {
+            $tids = implode(', ', $tids);
+            $movedids = implode("', '", $movedids);
+            $db->query("DELETE FROM ".X_PREFIX."threads WHERE tid IN ($tids) OR closed IN ('$movedids')");
+            $db->query("DELETE FROM ".X_PREFIX."favorites WHERE tid IN ($tids)");
+            $db->query("DELETE FROM d, r, v "
+                     . "USING ".X_PREFIX."vote_desc AS d "
+                     . "LEFT JOIN ".X_PREFIX."vote_results AS r ON r.vote_id = d.vote_id "
+                     . "LEFT JOIN ".X_PREFIX."vote_voters AS v  ON v.vote_id = d.vote_id "
+                     . "WHERE d.topic_id IN ($tids)");
+        }
+        echo "<p align=\"center\">Deleted ...</br>";
     }
 }
 
@@ -2013,6 +2039,7 @@ if ($action == "upgrade") {
     }
 
     if (onSubmit('upgradesubmit')) {
+        request_secure('insertrawsql', '', X_NONCE_FORM_EXP, FALSE);
         $upgrade = postedVar('upgrade', '', FALSE, FALSE);
         if (isset($_FILES['sql_file'])) {
             require('include/attach.inc.php');
@@ -2101,6 +2128,7 @@ if ($action == "upgrade") {
         <tr bgcolor="<?php echo $altbg2?>">
         <td align="center">
         <form method="post" action="cp.php?action=upgrade" enctype="multipart/form-data">
+        <input type="hidden" name="token" value="<?php echo nonce_create('insertrawsql'); ?>" />
         <table cellspacing="0" cellpadding="0" border="0" width="550" align="center">
         <tr>
         <td bgcolor="<?php echo $bordercolor?>">
