+++ cp2.php
@@ -234,10 +234,12 @@ if ($action == 'restrictions') {
 if ($action == 'lang') {
     if (noSubmit('importsubmit') And noSubmit('edit') And noSubmit('editsubmit') And noSubmit('detail') And noSubmit('deletesubmit')) {
         // Default screen: Language List, Options to Install, Uninstall, and Export.
+        $langnonce = nonce_create('massedtrnsls');
         ?>
         <tr bgcolor="<?php echo $altbg2?>">
         <td align="center">
         <form method="post" action="cp2.php?action=lang" name="theme_main">
+        <input type="hidden" name="token" value="<?php echo $langnonce; ?>" />
         <table cellspacing="0" cellpadding="0" border="0" width="500" align="center">
         <tr>
         <td bgcolor="<?php echo $bordercolor?>">
@@ -334,6 +336,7 @@ if ($action == 'lang') {
 
         <br />
         <form method="post" action="cp2.php?action=lang" enctype="multipart/form-data">
+        <input type="hidden" name="token" value="<?php echo $langnonce; ?>" />
         <table cellspacing="0" cellpadding="0" border="0" width="500" align="center">
         <tr>
         <td bgcolor="<?php echo $bordercolor?>">
@@ -359,6 +362,8 @@ if ($action == 'lang') {
     }
 
     if (onSubmit('importsubmit') && isset($_FILES['themefile']['tmp_name'])) { // Handle upload of new translation file.
+
+        request_secure('massedtrnsls', '', X_NONCE_FORM_EXP, FALSE);
 
         // Retrieve uploaded file
         require('include/attach.inc.php');
@@ -407,10 +412,12 @@ if ($action == 'lang') {
             $value = '';
         }
 
+        $key = template_key('edphrz', $phraseid);
         ?>
         <tr bgcolor="<?php echo $altbg2?>">
         <td align="center">
         <form method="post" action="cp2.php?action=lang">
+        <input type="hidden" name="token" value="<?php echo nonce_create($key); ?>" />
         <table cellspacing="0" cellpadding="0" border="0" width="550" align="center">
         <tr>
         <td bgcolor="<?php echo $bordercolor?>">
@@ -449,6 +456,8 @@ if ($action == 'lang') {
 
         $phraseid = getInt('phraseid', 'p');
         $newvalue = postedVar('templatenew', '', FALSE); // HTML is always allowed in translations.
+
+        request_secure('edphrz', $phraseid, X_NONCE_FORM_EXP, FALSE);
 
         if (!setLangValue($phraseid, $newvalue)) {
             error($lang['generic_missing'], FALSE);
@@ -519,6 +528,7 @@ if ($action == 'lang') {
     }
 
     if (onSubmit('deletesubmit')) {
+        request_secure('massedtrnsls', '', X_NONCE_FORM_EXP, FALSE);
         $theme_delete = postedArray('lang_delete', 'int');
         $result = $db->query("SELECT langid FROM ".X_PREFIX."lang_base WHERE devname='$langfile' OR devname='{$SETTINGS['langfile']}'");
         $lockIDs = array();
@@ -545,11 +555,14 @@ if ($action == 'themes') {
     $single_int = getInt('single');
     $newtheme = postedVar('newtheme');
 
+    $themenonce = nonce_create('massedthemes');
+
     if (noSubmit('themesubmit') && $single_str == '' && noSubmit('importsubmit')) {
         ?>
         <tr bgcolor="<?php echo $altbg2?>">
         <td>
         <form method="post" action="cp2.php?action=themes" name="theme_main">
+        <input type="hidden" name="token" value="<?php echo $themenonce; ?>" />
         <table cellspacing="0" cellpadding="0" border="0" width="500" align="center">
         <tr>
         <td bgcolor="<?php echo $bordercolor?>">
@@ -630,6 +643,7 @@ if ($action == 'themes') {
         </form>
         <br />
         <form method="post" action="cp2.php?action=themes" enctype="multipart/form-data">
+        <input type="hidden" name="token" value="<?php echo $themenonce; ?>" />
         <table cellspacing="0" cellpadding="0" border="0" width="500" align="center">
         <tr>
         <td bgcolor="<?php echo $bordercolor?>">
@@ -655,6 +669,7 @@ if ($action == 'themes') {
     }
 
     if (onSubmit('importsubmit') && isset($_FILES['themefile']['tmp_name'])) {
+        request_secure('massedthemes', '', X_NONCE_FORM_EXP, FALSE);
         if (!is_uploaded_file($_FILES['themefile']['tmp_name'])) {
             error($lang['textthemeimportfail'], FALSE);
         }
@@ -692,6 +707,7 @@ if ($action == 'themes') {
         }
         echo '</td></tr>';
     } else if (onSubmit('themesubmit')) {
+        request_secure('massedthemes', '', X_NONCE_FORM_EXP, FALSE);
         $theme_delete = postedArray('theme_delete', 'int');
         $theme_name = postedArray('theme_name', 'string', 'javascript', TRUE, TRUE, TRUE);
 
@@ -721,10 +737,12 @@ if ($action == 'themes') {
         $query = $db->query("SELECT * FROM ".X_PREFIX."themes WHERE themeid='$single_int'");
         $themestuff = $db->fetch_array($query);
         $db->free_result($query);
+        $key = template_key('theme', $single_int);
         ?>
         <tr bgcolor="<?php echo $altbg2?>">
         <td>
         <form method="post" action="cp2.php?action=themes&amp;single=submit">
+        <input type="hidden" name="token" value="<?php echo nonce_create($key); ?>" />
         <table cellspacing="0" cellpadding="0" border="0" width="93%" align="center">
         <tr>
         <td bgcolor="<?php echo $bordercolor?>">
@@ -845,6 +863,7 @@ if ($action == 'themes') {
         <tr bgcolor="<?php echo $altbg2?>">
         <td align="center">
         <form method="post" action="cp2.php?action=themes&amp;single=submit">
+        <input type="hidden" name="token" value="<?php echo nonce_create('makenewtheme'); ?>" />
         <table cellspacing="0" cellpadding="0" border="0" width="93%" align="center">
         <tr>
         <td bgcolor="<?php echo $bordercolor?>">
@@ -949,6 +968,8 @@ if ($action == 'themes') {
         </tr>
         <?php
     } else if ($single_str == "submit" && !$newtheme) {
+        $orig = formInt('orig');
+        request_secure('theme', $orig, X_NONCE_FORM_EXP, FALSE);
         $namenew = postedVar('namenew');
         $bgcolornew = postedVar('bgcolornew');
         $altbg1new = postedVar('altbg1new');
@@ -975,6 +996,7 @@ if ($action == 'themes') {
         $db->query("UPDATE ".X_PREFIX."themes SET name='$namenew', bgcolor='$bgcolornew', altbg1='$altbg1new', altbg2='$altbg2new', link='$linknew', bordercolor='$bordercolornew', header='$headernew', headertext='$headertextnew', top='$topnew', catcolor='$catcolornew', tabletext='$tabletextnew', text='$textnew', borderwidth='$borderwidthnew', tablewidth='$tablewidthnew', tablespace='$tablespacenew', fontsize='$fsizenew', font='$fnew', boardimg='$boardlogonew', imgdir='$imgdirnew', smdir='$smdirnew', cattext='$cattextnew', admdir='$admdirnew' WHERE themeid='$orig'");
         echo '<tr bgcolor="'.$altbg2.'" class="ctrtablerow"><td>'.$lang['themeupdate'].'</td></tr>';
     } else if ($single_str == "submit" && $newtheme) {
+        request_secure('makenewtheme', '', X_NONCE_FORM_EXP, FALSE);
         $namenew = postedVar('namenew');
         $bgcolornew = postedVar('bgcolornew');
         $altbg1new = postedVar('altbg1new');
@@ -1296,6 +1318,7 @@ if ($action == "ranks") {
         <tr bgcolor="<?php echo $altbg2?>">
         <td align="center">
         <form method="post" action="cp2.php?action=ranks">
+        <input type="hidden" name="token" value="<?php echo nonce_create('editusrranks'); ?>" />
         <table cellspacing="0" cellpadding="0" border="0" width="650" align="center">
         <tr>
         <td bgcolor="<?php echo $bordercolor?>">
@@ -1360,6 +1383,7 @@ if ($action == "ranks") {
         </tr>
         <?php
     } else {
+        request_secure('editusrranks', '', X_NONCE_FORM_EXP, FALSE);
         $id = postedArray('id', 'int');
         $delete = postedArray('delete', 'int');
         $title = postedArray('title', 'string', '', FALSE);
@@ -1412,6 +1436,7 @@ if ($action == "newsletter") {
         <td>
         <form method="post" action="cp2.php?action=newsletter">
         <table cellspacing="0" cellpadding="0" border="0" width="550" align="center">
+        <input type="hidden" name="token" value="<?php echo nonce_create('sendnewslttr'); ?>" />
         <tr>
         <td bgcolor="<?php echo $bordercolor?>">
         <table border="0" cellspacing="<?php echo $THEME['borderwidth']?>" cellpadding="<?php echo $tablespace?>" width="100%">
@@ -1467,6 +1492,7 @@ if ($action == "newsletter") {
         </tr>
         <?php
     } else {
+        request_secure('sendnewslttr', '', X_NONCE_FORM_EXP, FALSE);
         @set_time_limit(0);
         $newssubject = postedVar('newssubject');
         $newsmessage = postedVar('newsmessage');
@@ -1545,6 +1571,7 @@ if ($action == "prune") {
         <tr bgcolor="<?php echo $altbg2?>">
         <td align="center">
         <form method="post" action="cp2.php?action=prune">
+        <input type="hidden" name="token" value="<?php echo nonce_create('admmassprune'); ?>" />
         <table cellspacing="0" cellpadding="0" border="0" width="550">
         <tr>
         <td bgcolor="<?php echo $bordercolor?>">
@@ -1582,7 +1609,7 @@ if ($action == "prune") {
         <input type="checkbox" name="pruneByPosts[check]" value="1" />
         </td>
         <td>
-        <select name="pruneBy[posts][type]">
+        <select name="pruneByPosts[type]">
         <option value="more"><?php echo $lang['prunemorethan']?></option>
         <option value="is"><?php echo $lang['pruneexactly']?></option>
         <option value="less"><?php echo $lang['prunelessthan']?></option>
@@ -1648,6 +1675,7 @@ if ($action == "prune") {
         </tr>
         <?php
     } else {
+        request_secure('admmassprune', '', X_NONCE_FORM_EXP, FALSE);
         $pruneByDate = postedArray('pruneByDate');
         $pruneByPosts = postedArray('pruneByPosts');
         $pruneFrom = postedVar('pruneFrom', '', FALSE, FALSE);
@@ -1831,6 +1859,7 @@ if ($action == "templates") {
         <tr bgcolor="<?php echo $altbg2?>">
         <td align="center">
         <form method="post" action="cp2.php?action=templates">
+        <input type="hidden" name="token" value="<?php echo nonce_create('rsttemplates'); ?>" />
         <table cellspacing="0" cellpadding="0" border="0" width="550" align="center">
         <tr>
         <td bgcolor="<?php echo $bordercolor?>">
@@ -1855,6 +1884,7 @@ if ($action == "templates") {
     }
 
     if (onSubmit('restoresubmit')) {
+        request_secure('rsttemplates', '', X_NONCE_AYS_EXP, FALSE);
         if (!file_exists('./templates.xmb')) {
             error($lang['no_templates'], false, '</td></tr></table></td></tr></table><br />');
         }
@@ -1886,10 +1916,12 @@ if ($action == "templates") {
             error($lang['selecttemplate'], false, '</td></tr></table></td></tr></table><br />');
         }
         $tid = formInt('tid');
+        $key = template_key('tmplt', $tid);
         ?>
         <tr bgcolor="<?php echo $altbg2?>">
         <td align="center">
         <form method="post" action="cp2.php?action=templates&amp;tid=<?php echo $tid?>">
+        <input type="hidden" name="token" value="<?php echo nonce_create($key); ?>" />
         <table cellspacing="0" cellpadding="0" border="0" width="550" align="center">
         <tr>
         <td bgcolor="<?php echo $bordercolor?>">
@@ -1926,6 +1958,7 @@ if ($action == "templates") {
 
     if (onSubmit('editsubmit')) {
         $tid = postedVar('tid', '', FALSE, FALSE, FALSE, 'g');
+        request_secure('tmplt', $tid, X_NONCE_FORM_EXP, FALSE);
         $namenew = postedVar('namenew');
         //Templates are historically double-slashed.
         $templatenew = $db->escape(addslashes(postedVar('templatenew', '', FALSE, FALSE)));
@@ -1954,10 +1987,12 @@ if ($action == "templates") {
             error($lang['selecttemplate'], false, '</td></tr></table></td></tr></table><br />');
         }
         $tid = getInt('tid', 'r');
+        $key = template_key('dtmpl', $tid);
         ?>
         <tr bgcolor="<?php echo $altbg2?>">
         <td align="center">
         <form method="post" action="cp2.php?action=templates&amp;tid=<?php echo $tid?>">
+        <input type="hidden" name="token" value="<?php echo nonce_create($key); ?>" />
         <table cellspacing="0" cellpadding="0" border="0" width="550" align="center">
         <tr>
         <td bgcolor="<?php echo $bordercolor?>">
@@ -1983,6 +2018,7 @@ if ($action == "templates") {
 
     if (onSubmit('deletesubmit')) {
         $tid = getInt('tid', 'r');
+        request_secure('dtmpl', $tid, X_NONCE_AYS_EXP, FALSE);
         $db->query("DELETE FROM ".X_PREFIX."templates WHERE id=$tid");
         echo '<tr bgcolor="'.$altbg2.'" class="ctrtablerow"><td>'.$lang['templatesdelete'].'</td></tr>';
         redirect($full_url.'cp2.php?action=templates', 2, X_REDIRECT_JS);
@@ -1990,10 +2026,12 @@ if ($action == "templates") {
 
     if (onSubmit('new')) {
         $newtemplatename = postedVar('newtemplatename', 'javascript', TRUE, FALSE, TRUE);
+        $key = template_key('tmplt', 'new');
         ?>
         <tr bgcolor="<?php echo $altbg2?>">
         <td align="center">
         <form method="post" action="cp2.php?action=templates&amp;tid=new">
+        <input type="hidden" name="token" value="<?php echo nonce_create($key); ?>" />
         <table cellspacing="0" cellpadding="0" border="0" width="550" align="center">
         <tr>
         <td bgcolor="<?php echo $bordercolor?>">
@@ -2094,6 +2132,7 @@ if ($action == "attachments") {
         <tr bgcolor="<?php echo $altbg2?>">
         <td align="center">
         <form method="post" action="cp2.php?action=attachments">
+        <input type="hidden" name="token" value="<?php echo nonce_create('massedattach'); ?>" />
         <table cellspacing="0" cellpadding="0" border="0" width="93%" align="center">
         <tr>
         <td bgcolor="<?php echo $bordercolor?>">
@@ -2264,6 +2303,7 @@ if ($action == "attachments") {
     }
 
     if (onSubmit('deletesubmit')) {
+        request_secure('massedattach', '', X_NONCE_FORM_EXP, FALSE);
         require('include/attach.inc.php');
         $filelist = array();
         foreach($_POST as $postedname => $rawvalue) {
@@ -2530,11 +2570,24 @@ if ($action == "cplog") {
 }
 
 if ($action == "delete_attachment") {
-    require('include/attach.inc.php');
     $aid = getInt('aid');
     $pid = getInt('pid');
-    deleteAttachment($aid, $pid);
-    echo "<p align=\"center\">Deleted ...</br>";
+    if (noSubmit('yessubmit')) {
+        $key = template_key('delat', $aid);
+        ?>
+        <tr bgcolor="<?php echo $altbg2; ?>" class="ctrtablerow"><td>Are you sure you want to delete this attachment?<br />
+        <form action="cp2.php?action=delete_attachment&amp;aid=<?php echo $aid; ?>&amp;pid=<?php echo $pid; ?>" method="post">
+          <input type="hidden" name="token" value="<?php echo nonce_create($key); ?>" />
+          <input type="submit" name="yessubmit" value="<?php echo $lang['textyes']; ?>" /> -
+          <input type="submit" name="yessubmit" value="<?php echo $lang['textno']; ?>" />
+        </form></td></tr>
+        <?php
+    } elseif ($lang['textyes'] == $yessubmit) {
+        request_secure('delat', $aid, X_NONCE_AYS_EXP, FALSE);
+        require('include/attach.inc.php');
+        deleteAttachment($aid, $pid);
+        echo "<p align=\"center\">Deleted ...</br>";
+    }
 }
 
 if ($action == "movetodb_attachment") {
